Facebook
Refactor project for structured foldering. Add ability to get href hostname.
